[fix/#74] repository 통일
GroupBuy -> Party
