+++ app/src/main/java/com/poti/android/data/di/ServiceModule.kt
@@ -2,7 +2,6 @@ package com.poti.android.data.di
 
 import com.poti.android.data.remote.service.ArtistService
 import com.poti.android.data.remote.service.AuthService
-import com.poti.android.data.remote.service.GroupBuyService
 import com.poti.android.data.remote.service.HomeService
 import com.poti.android.data.remote.service.ImageService
 import com.poti.android.data.remote.service.PartyService
@@ -46,9 +45,4 @@ object ServiceModule {
     @Singleton
     fun provideHomeService(retrofit: Retrofit): HomeService =
         retrofit.create(HomeService::class.java)
-
-    @Provides
-    @Singleton
-    fun provideGroupBuyService(retrofit: Retrofit): GroupBuyService =
-        retrofit.create(GroupBuyService::class.java)
 }
