+++ app/src/main/java/com/poti/android/data/remote/datasource/PartyRemoteDataSource.kt
@@ -3,6 +3,7 @@ package com.poti.android.data.remote.datasource
 import com.poti.android.core.network.model.BaseResponse
 import com.poti.android.data.remote.dto.request.party.CreatePartyRequestDto
 import com.poti.android.data.remote.dto.response.artist.ArtistSearchListResponseDto
+import com.poti.android.data.remote.dto.response.history.GroupBuyPostSaleDto
 import com.poti.android.data.remote.dto.response.party.CreatePartyResponseDto
 import com.poti.android.data.remote.dto.response.party.ProductSearchResponseDto
 import com.poti.android.data.remote.dto.response.party.ShippingOptionResponseDto
@@ -26,4 +27,7 @@ class PartyRemoteDataSource @Inject constructor(
 
     suspend fun getShippingOptions(): BaseResponse<List<ShippingOptionResponseDto>> =
         partyService.getShippingOptions()
+
+    suspend fun getPostSale(postId: Long): BaseResponse<GroupBuyPostSaleDto> =
+        partyService.getPostSale(postId)
 }
