+++ app/src/main/java/com/poti/android/data/remote/service/PartyService.kt
@@ -3,12 +3,14 @@ package com.poti.android.data.remote.service
 import com.poti.android.core.network.model.BaseResponse
 import com.poti.android.data.remote.dto.request.party.CreatePartyRequestDto
 import com.poti.android.data.remote.dto.response.artist.ArtistSearchListResponseDto
+import com.poti.android.data.remote.dto.response.history.GroupBuyPostSaleDto
 import com.poti.android.data.remote.dto.response.party.CreatePartyResponseDto
 import com.poti.android.data.remote.dto.response.party.ProductSearchResponseDto
 import com.poti.android.data.remote.dto.response.party.ShippingOptionResponseDto
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
+import retrofit2.http.Path
 import retrofit2.http.Query
 
 interface PartyService {
@@ -30,4 +32,9 @@ interface PartyService {
 
     @GET("/api/v1/shippings")
     suspend fun getShippingOptions(): BaseResponse<List<ShippingOptionResponseDto>>
+
+    @GET("/api/v1/posts/sale/{postId}")
+    suspend fun getPostSale(
+        @Path("postId") postId: Long,
+    ): BaseResponse<GroupBuyPostSaleDto>
 }
