--- app/src/main/java/com/poti/android/domain/repository/GroupBuyRepository.kt
@@ -1,7 +1,0 @@
-package com.poti.android.domain.repository
-
-import com.poti.android.domain.model.history.RecruiterDetail
-
-interface GroupBuyRepository {
-    suspend fun getPostSale(postId: Long): Result<RecruiterDetail>
-}
